requirements.txt
PySide6>=6.5,<7
pyserial>=3.5,<4
chipshouter>=1.0
